//#EJC20220101: Corrección tipo fecha int to Long en clsBeP_DESCUENTO

+++ app/src/main/java/com/dtsgt/classes/clsDocument.java
@@ -25,7 +25,7 @@ public class clsDocument {
 	public boolean docfactura,docrecibo,docanul,docpedido,docdevolucion,doccanastabod;
 	public boolean docdesglose,pass,facturaflag,banderafel,propfija,impresionorden;
 	public boolean parallevar,factsinpropina,modorest;
-	public Long ffecha;
+	public long ffecha;
     public int pendiente,diacred,pagoefectivo;
 	//#EJC20210705
 	public String TipoCredito, NoAutorizacion;
@@ -915,12 +915,12 @@ public class clsDocument {
 	}
 	
 	public String sfecha(long f) {
-		int vy,vm,vd;
+	    long vy,vm,vd;
 		String s;
 		
-		vy=(int) f/100000000;f=f % 100000000;
-		vm=(int) f/1000000;f=f % 1000000;
-		vd=(int) f/10000;f=f % 10000;
+		vy=(long) f/100000000;f=f % 100000000;
+		vm=(long) f/1000000;f=f % 1000000;
+		vd=(long) f/10000;f=f % 10000;
 		
 		s="";
 		if (vd>9) { s=s+String.valueOf(vd)+"-";} else {s=s+"0"+String.valueOf(vd)+"-";}
